+++ compute.go
@@ -81,6 +81,10 @@ func figContainersManager(ctx Config, com Component) func(gfx *dot.Graph) bool {
 			return false
 		}
 
+		if lab := strings.TrimSpace(com.Label); len(lab) == 0 {
+			com.Label = "Containers\nEngine"
+		}
+
 		if fc := strings.TrimSpace(com.FontColor); len(fc) == 0 {
 			com.FontColor = "#fafafaff"
 		}
@@ -92,7 +96,7 @@ func figContainersManager(ctx Config, com Component) func(gfx *dot.Graph) bool {
 		cl := cluster.New(gfx, com.ID, cluster.BottomTop(ctx.bottomTop), cluster.Label(com.Impl))
 
 		node.New(cl, com.ID,
-			node.Label("<b>Containers<br/>Engine</b>", true),
+			node.Label(com.Label, false),
 			node.FontSize(7),
 			node.FontColor(com.FontColor),
 			node.FillColor(com.FillColor),